requirements.txt
python-telegram-bot
numpy
opencv-python
python-dotenv
PyYAML